bringing Class 2 into the Class 1 Blueprint
Changes to text and descriptions to bring class 2 into the class 1
blueprint. the only technical changes that needed to be adjusted were
the collections referencing the old BIGIP A management and the old
nodes which are no longer used in place of the containers. Module 5
will come later this week (the new GIT Commit work). I wanted a good
baseline before introducing that