class2 fixes (#170)
* Class2 errata

mod2lab1:
and a automated Test Framework. > and an automated test framework.
closed an relaunched. > closed and relaunched.
To utilize the helper functions the framework includes > To utilize the helper functions included in the framework

mod2lab2:
fix image for selecting the environment

mod2lab4:
If you need to launch the container once more, follow the instructions in this previous lab and log in as snops.

mod2lab5:
to the the 1_Authenticate > to the 1_Authenticate >

mod3lab2:
#. The above f5-newman-wrapper  <--fix indentation
Delete any virtual servers or Pools that appear in the GUI.  <--add to be more explicit

mod3lab3:
f5-newman-build-3 contains > f5-newman-build-4 contains

mod4mod4:
 if it’s not please refer to Class 2 Module 2 on starting the service. >  if it’s not please refer to Class 2 Module 1 Lab 2 on starting the service.

mod4lab1:
minor edits to Slack explanation

mod4lab2:
Once the Job is saved, <---fix indentation
Highlight over the third Stage to prompt for the Approval   <---fix indentation
Repeat steps 2 & 3 of the last module,   <--Repeats steps 2 & 3 of the previous task to create 2 new Jenkins jobs,

mod4lab3
Destruction of an Application Services is > Destruction of Application Services is
Repeat steps 2 & 3 of the first module > Repeat steps 2 & 3 of task 1 in the previous lab
if you have extra time > If you have extra time

* trying to fix indentation and numbered lists